# [ArrowStringArray] PERF: str.partition object fallback