```
       before           after         ratio
     [40482273]       [5d1bc3e0]
     <master>         <wrap-result>
-      59.6±0.5ms       36.2±0.6ms     0.61  strings.Methods.time_partition('arrow_string')
-      57.8±0.5ms       34.9±0.2ms     0.60  strings.Methods.time_rpartition('arrow_string')
-      54.5±0.3ms       32.0±0.4ms     0.59  strings.Methods.time_partition('str')
-      53.3±0.2ms       30.9±0.5ms     0.58  strings.Methods.time_rpartition('str')
-      49.0±0.5ms       26.4±0.3ms     0.54  strings.Methods.time_partition('string')
-      48.2±0.2ms       25.7±0.4ms     0.53  strings.Methods.time_rpartition('string')

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE INCREASED.
```